third_party/WebKit/LayoutTests/http/tests/inspector/persistence/persistence-tabbed-editor-keeps-selected-tab.html

<html>
<head>
<script src='../inspector-test.js'></script>
<script src='../debugger-test.js'></script>
<script src='../isolated-filesystem-test.js'></script>
<script src='./persistence-test.js'></script>
<script src='./resources/foo.js'></script>
<script>

async function test() {
  var testMapping = BindingsTestRunner.initializeTestMapping();
  var fs = new BindingsTestRunner.TestFileSystem('file:///var/www');
  var fsEntry = BindingsTestRunner.addFooJSFile(fs);
  fs.root.addFile('bar.js', 'window.bar = ()=>\'bar\';');
  await fs.reportCreatedPromise();

  var fsSourceCode = await TestRunner.waitForUISourceCode('foo.js', Workspace.projectTypes.FileSystem);
  var networkSourceCode = await TestRunner.waitForUISourceCode('foo.js', Workspace.projectTypes.Network);
  var barSourceCode = await TestRunner.waitForUISourceCode('bar.js');
  UI.panels.sources.showUISourceCode(barSourceCode, 0, 0);
  UI.panels.sources.showUISourceCode(networkSourceCode, 0, 0);
  // Open and select file system tab. Selection should stay here.
  UI.panels.sources.showUISourceCode(fsSourceCode, 0, 0);

  dumpTabs('Opened tabs before persistence binding:');
  testMapping.addBinding('foo.js');
  await BindingsTestRunner.waitForBinding('foo.js');
  dumpTabs('\nOpened tabs after persistence binding:');
  TestRunner.completeTest();

  function dumpTabs(title) {
    var tabbedPane = UI.panels.sources._sourcesView._editorContainer._tabbedPane;
    var tabs = tabbedPane._tabs;
    TestRunner.addResult(title);
    for (var i = 0; i < tabs.length; ++i) {
      var text = (i + 1) + ': ';
      text += tabs[i].title;
      if (tabs[i] === tabbedPane._currentTab)
        text += ' [selected]';
      TestRunner.addResult('    ' + text);
    }
  }
};
</script>
</head>
<body onload='runTest()'>
<p>Verify that tab keeps selected as the persistence binding comes in.</p>
</body>
</html>
